ben-sandbox.py
"""
Example usage of the Qualtrics API module
This script demonstrates how to create surveys with various question types
"""

import os
import sys
from pathlib import Path

# Add parent directory to path for imports
sys.path.insert(0, str(Path(__file__).parent.parent))

from dotenv import load_dotenv
from qualtrics_sdk import QualtricsAPI

# Load credentials from .env file
load_dotenv()

API_TOKEN = os.getenv('QUALTRICS_API_TOKEN')
DATA_CENTER = os.getenv('QUALTRICS_DATA_CENTER')

if not API_TOKEN or not DATA_CENTER:
    raise ValueError("Please set QUALTRICS_API_TOKEN and QUALTRICS_DATA_CENTER in .env file")

# Initialize the API client
api = QualtricsAPI(api_token=API_TOKEN, data_center=DATA_CENTER)


api.get_survey("SV_83cKBS1RcEBqVMi")